Education-Management-System/backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
PyJWT==2.1.0
bcrypt==3.2.0
python-dateutil==2.8.2
